refactor: - migrated all update code to the bridge - migration to streaming
