+++ backend/server/cmd/router.go
@@ -19,7 +19,6 @@ import (
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/control"
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/filebrowser"
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/network"
-	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/updates"
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/wireguard"
 	"github.com/mordilloSan/LinuxIO/backend/server/web"
 )
@@ -58,7 +57,6 @@ func BuildRouter(cfg Config, sm *session.Manager) *gin.Engine {
 	})
 
 	// Protected endpoints:
-	updates.RegisterUpdateRoutes(r.Group("/updates", sm.RequireSession()))
 	network.RegisterNetworkRoutes(r.Group("/network", sm.RequireSession()))
 	wireguard.RegisterWireguardRoutes(r.Group("/wireguard", sm.RequireSession()))
 	config.RegisterThemeRoutes(r.Group("/theme", sm.RequireSession()))
